requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.17.0
